[server] 리눅스 빌드에러수정 (Windows API 종속된 부분 일단 제거)

+++ TCP_SkyrimServer_ProjectFile/customVector.cpp
@@ -61,7 +61,7 @@ void vector::normalize(vector *pOut, const vector *pV)
 	if (pV == nullptr)
 		return ;
 
-	int sqrtSum = sqrt(pV->x) + sqrt(pV->y) + sqrt(pV->z);
+	float sqrtSum = sqrt(pV->x) + sqrt(pV->y) + sqrt(pV->z);
 	pOut->x = (pV->x)/sqrtSum;
 	pOut->y = (pV->y)/sqrtSum;
 	pOut->z = (pV->z)/sqrtSum;
